Add capturing audio preview.
Add capturing of the preview input to the Cognitive speech service. This
ensures a few things like our assumptions that the BufferedwaveProviders
are working as we intended (are being written to, can be written from
etC).

Not that we do not want to "read fully" option enabled since this inserts
gaps in the audio stream that makes the voice recognition harder.