Add a Speech SDK BufferedWaveProvider StreamReader
Add a class that implements the PullAudioInputStreamCallback class and
reads from a NAudio BufferedWaveProvider. This will allow our Speech SDK
recognizers to be able to read data from NAudio buffers.

In order to be able to use the Speech SDK we also need to upgrade from
targetting .Net 4.6 to .Net 4.7