Handle combined numbers and dashless callsign
Sometimes ths peech recognition combines the spoken numbers "one three"
and the like to 13, even though "one one" is correctly transcribed as
"one one".

We will also make the dash character in the nickname optional. This
means the bot will be able to handle call-signs with the following
format

* DOLT 1-1 | Player
* DOLT 11 | Player

Which covers two common formats.