Slow tranmission rate of audio data
Because SRS was built around real-time human <-> human communication;
normal SRS clients have an audio buffer of 3 seconds. However the bot
stores an entire response before streaming it as fast as it can to
clients.

This means we can send 8 seconds of audio data in a fraction of that
time which results in the client buffers getting full and the end of the
bot audio getting chopped off.

Therefore we will slow the rate at which we send the audio-data. Slowing
it by 10-20ms for every 40ms of audio data we send is slow enough to not
overflow the buffers without causing network related breaks in audio.