Add airframe type information to responses
May not be 100% true-to-life since it magically knows what the airframe
is but players don't like it right now because they could be running
away from an Su-25T instead of a fighter.

Also translate the DCS aircraft type to NATO reporting names where
possible (or the Western name, i.e. F-14 = Tomcat) of common Opfor
airfact on Hoggit servers where this bot is curently running.

Eventually we can replce the switch statement with a config file lookup
(YAML or JSON or something).