Wrap warning radius checkers in try block
This is a blunt fix to stop the entire bot crashing when something goes
wrong checking the warning radius. Will need to consider possible
scenarios on better ways to handle this.

Currently it means no warning will go out but since the adding of the
contact's ID is the very last thing done now it should mean the warning
goes out 5 seconds later on the next check assuming nothing goes wrong
there as well.