Initial Declare implementation
Initial implementation of a declare call. Probably only suitable for
short range declares (i.e. from A-10s and Ground Pounders).

Bearing is a required value and if distance is not given then it is
assumed to be one mile. The bot will check all contacts within 1
nautical mile of the point specified by the bearing and distance.

This is good enough for short range declare calls but probably not
accurate enough for long-range declare calls.