Use the DB connection properly to fix concurrency.
Use the DB connection properly (i.e. a new one each time and let the
pool figure it out). Up to now we were always using one connection due
to not understanding fully how it worked. This was fine with very low
traffic but leads to concurrency issues with many plaers.

We will also delegate logging of SQL to Npgsql itself, it uses debug by
default instead of trace but this is not a showstopper and we will leave
it like this for the moment. Maybe having trace logging is not needed
right now