Add special handling for Caucasus bearings.
The Caucasus map is special in that it does not match up to reality.
this means that if we use real-life calculations such as Haversine and
Magnetic Variation libraries we get the wrong results.

This is one reason why declare calls are failing so much, because the
bearings do not match up. It doesn't really matter if a Bogey dope call
is incorrect by 5 degrees but it does for declares.

Therefore add some special handling for the Caucasus map. Caucasus True
heading is actualy real-life Magnetic so we need to double it up in our
calculations. Since the map also doesn't reflect reality we need to add
a fudge factor. I am trying to make the bot as close to LotATC as
possible in terms of given headings.

It is still not perfect but as much as I am willing to do on this,
Declare calls should be more accurate now.

Many thanks to DArt (LotATC developer) for revealing the...
special nature of the Caucasus map.