Rewrite Voice Recognition code
Rewrite Voice Recognition code to cater more to OverlordBots infrequent
use.

Instead of doing continuous recognition we will switch to 1-shot
recognition which is more designed for commands anyway. The advantage of
this is that the connection to the MS servers doesn't stay open for 30
seconds but finishes as soon as the transmission finishes.

This reduction reduces the time window in which we might break our
connection limit which makes supporting an extra frequency (on the same
or a different server) more likely to work.