Get aircraft near Anapa.
Get all aircraft within 10nm of Anapa that are less than 3000ft high.
This is the first step to narrowing down the list of aircraft that might
be subject to atc.

If we didn't do this first step then we would run into issues in the
future where we would be checking the position of N aircraft against M
navigation points which any Big O notation person will tell you is bad
juju.