Determine aircraft is within a navigation point
Determine if an aircraft is within a navigation point. This will be
needed to be able to track aircraft through a state machine based on
position.

As part of this have created some navigation points for Anapa and Kras
Center.

The NetTopologySuite library, while allowing for npgsql integration, has
no calculations or anything (In fact it is a very dumb library).
Therefore we should switch to the Geo library for all the interesting
stuff. NetTopologySuite's objects should not survive past the GameState
class.