Move aircraft definitions to JSON file
Move aircraft definition information to a separate JSON file and add
more entries with NATO reporting names.

The JSON is a list of hashes which makes it easy to use with JSON.Net
and also allows for future-proofing by making it easy to expand with
more attributes in future (e.g. maybe a threat rating to help prioritise
threat calls).

Co-authored-by: ugene143 <e12.ugene@gmail.com>