Update to Directed Graph for Traffic flow
Update to a directed graph and redo all the pathing to enable traffic
flow control on more complicated airfields.

Also fill in the spawn point areas. In order to make things less painful
for future mapping we may want to remove the Polygon and replace with a
"to nearest point" function to determine where aircraft are.