Extract ATC functions to separate project
Extract the ATC functions to a separate project which makes them easier
to test and package separately as a Nuget in future if there is a demand
for it.

While we are at it we will also add tests to make sure the taxi
instructions are as we expect them to be.