Switch to taxipoints with positions
Switch to recording all Taxipoints with lat/long instead of defining
polygons for the parking areas.

The advantages are that we can now request taxi instructions from
basically all points on the airfield and it removes the requirements for
mapping out the parking spots.

The disadvantage is that we need to mark the position of all the taxi
points and that we need some extra points around the parking areas
sometimes to make sure planes are marked as being in the correct
location. This can get tricky when planes are in locations close to, but
not part of a taxi point.

For example Slot 40 on the East Apron is currently close to an Echo
Spots marker point than an East Apron point so it is being incorrectly
recorded. It doesn't matter right now because the taxi instructions are
the same but it is still something that needs to be fixed.

Have submitted a request for volunteers in the #scripting channel on the
hoggit server and on the ED forums
https://forums.eagle.ru/showthread.php?t=280261 . We will see if we can
get any help.

For marking the current spots in Anapa I did it manually as a proof of
concept