Add some basic visual editing
Add some basic visual editing features which allow you to create
taxiways between nodes at a fixed cost of 1. It tries to guess the name
but is pretty damn stupid.

The aim of this change is to remove the most painful part of adding the
taxiways which is all the copying and pasting and typing.