Special-Case the logic for picking active runways
Some graphs are setup to not follow the default wind of 090 so we will
special case these for the moment.

As part of this consolidate all the tests for taxiing to a runway in a
single file and use DynamicDataTest attributes to reduce the amount of
typing required. We will also add tests for the special cased airfields
to make sure they behave as expected.