Move more functionality into the Controllers
Move more functionality into the controller and out of the Speech
Recogniser. This is causing most controller tests to fail since we now
need to mock out the GameQuerier which we are not yet doing.

Some other test fixes also because I am not being that fussed about
atoimcity for a 1 man hobby project.