Fix 360 wrap around issue
Fix issue when wind heading is less than 090 which causes wrap-around
issues.

There is still an issue whereby if the wind is exactly 90 degrees to a
runway then the bot will send players to the closet end of said runway
instead of aligning all players. This is next on the list.