Move Data files
Move the Aircraft.json data file to top level Data directory so that it
sits in the same one as the Airfields datafiles.

Stop building the Airfields.json file as it is no longer in use but keep
it around as people use it to get coords for airfield mapping.