Massive refactoring
Up to now we have avoided doing large-scale refactoring becaus it
increases the difference in code that makes merging harder when it is
required.

However, in order to be able to send multiple transmissions at once, we
are going to have to make some major changes to how the SRS component is
coded so will lose that compatibility anyway.

Therefore go whole hog and do a lot of refactoring to clean up a lot of
longstanding isssues