Provide reactive runway crossing call
Provide a reactive call that will tell the pilot to cross the next
runway at their discretion as they approach it.

At the moment this is the only call. There is no "hold short" or other
intelligent handling yet.

Get rid of the initial call to cross runways in the initial
instructions, update tests and remove tests that are now hard to run