Add WayPoint and NavigationPoint
As well as TaxiPoints we will have WayPoints for inbound and outbound
aircraft still flying. Create a new NavigationPoint class as the parent
class for this and TaxiPoints so we can put them all on the same graph
while still being able to be specific about what points we want when
querying.