Add metrics
Add metrics for OverlordBot related to SRS and Radio usage as well as
some thread information to try and help debug a leak.

As part of this we are also including other changes made as a result of
seeing these metrics live. They should really be split out into other
changes but everything was so meshed together as I learned how to do
these metrics it is ending up in one commit.

Note that these metrics are not using the OpenTelemetry API but using
NewRelic SDK directly since the APIs are all experimental anyway and
this is the first thing I got working.