Don't initialize synthesizer on every call
Replace the per-call synthesizer with a permanent one.

This is the recommended practice but it keeps failing after a while of
non-usage with an HTTP_READ_ERROR, however a subsequent call works
without issues. Therefore we will hard-code an immediate second attempt
if we get a synthesis error and only return an error response if both
attempts fail.

Radio calls when the synthesizer has not "timed out" (for want of a
better word) will respond a *lot* faster (Half a second or more). Radio
calls that have to do the two attempts will still be about the same as
before, so overall we have a reduction in average radio call time.